Strip struct before encoding as JSON.
The protobuf struct generator doesn't include `@derive` module
attributes to setup a JSON encoder. This casuses errors when trying to
encode the structs. Since, by the point of encoding we have already
verified that the returned data is the expected struct we can trim off
the struct field and encode the data as a map.